Check Secure Message size before reading
Add range checks to Secure Message entry points before reading header
contents like this:

    themis_secure_message_hdr_t* message_hdr = (themis_secure_message_hdr_t*)encrypted_message;
    THEMIS_CHECK_PARAM(IS_THEMIS_SECURE_MESSAGE_ENCRYPTED(message_hdr->message_type));
    THEMIS_CHECK_PARAM(encrypted_message_length >= THEMIS_SECURE_MESSAGE_LENGTH(message_hdr));

If we don't then those lines will cause a buffer overrun when reading
from a buffer of length 1, for example.

I won't go down here on how parsing data like this breaks strict pointer
aliasing rules and might break on architectures with strict alignment
requirements. However, we MUST ensure that the buffer we read from has
at least enough for that "themis_secure_message_hdr_t" that we read.

Note that for in sign/verify mode the header is actually larger.
However, in this case we are basically interested in the first few bytes
of the Secure Message header which are the same size in all modes:

    sizeof(themis_secure_message_hdr_t) <= sizeof(themis_secure_signed_message_hdr_t)
    sizeof(themis_secure_message_hdr_t) <= sizeof(themis_secure_encrypted_message_hdr_t)

Therefore it's okay to check only for the length of the thing we are
casting to. Even in themis_secure_message_unwrap() which handles both
decryption and verification.