Fix a logic bug in signed message length check
Look closely at this condition. Very-very closely:

    themis_secure_signed_message_hdr_t* msg = (themis_secure_signed_message_hdr_t*)wrapped_message;
    if (((msg->message_hdr.message_type == THEMIS_SECURE_MESSAGE_RSA_SIGNED
          && soter_verify_get_alg_id(ctx->verify_ctx) != SOTER_SIGN_rsa_pss_pkcs8)
         || (msg->message_hdr.message_type == THEMIS_SECURE_MESSAGE_EC_SIGNED
             && soter_verify_get_alg_id(ctx->verify_ctx) != SOTER_SIGN_ecdsa_none_pkcs8))
        && (msg->message_hdr.message_length + msg->signature_length + sizeof(themis_secure_message_hdr_t)
            > wrapped_message_length)) {
        return THEMIS_INVALID_PARAMETER;
    }

You see a bug? Here, let me rewrite this mess a bit so that it becomes
more readable:

    if (
         (
           (message_signed_with_RSA   && verify_ctx_does_NOT_use_RSA)
           ||
           (message_signed_with_ECDSA && verify_ctx_does_NOT_use_ECDSA)
         )
         &&
         (wrapped_message_length < what_it_should_be_based_on_the_header)
       )
    {
        return THEMIS_INVALID_PARAMETER;
    }

Now you see it? Not really, I believe. However, note *what* we check for
in the first two conditions. They check for invalid combinations of the
verifier configuration and signature type in the message. That is,

    if (
         false_if_verifier_has_the_same_signature_type_as_in_the_message_
         which_is_of_course_exactly_that_for_all_valid_messages_actually
         &&
         (wrapped_message_length < what_it_should_be_based_on_the_header)
       )
    {
        return THEMIS_INVALID_PARAMETER;
    }

Now you see it? For *valid* messages, the length check is ignored
because the first part of the condition is always false.

The correct thing to do here is to check for all of those conditions:
if *either* of them is triggered, we should return an error.