Avoid integer overflow in the length check
And another one! What a plentiful harvest of bugs this function is.
Blame C for that.

Look closely at this sum expression. Note that "message_length" and
"signature_length" are both uint32_t. The next part is a sizeof()
expression which has size_t type, and "wrapped_message_length" is
also a size_t. So here we have

    uint32_t + uint32_t + size_t > size_t

The left part will be promoted to size_t so no problems, right? Wrong.
Addition operator has left associativity, meaning that the expression
above is computed as

    (uint32_t + uint32_t) + size_t > size_t

Both these uint32_t values are provided by the user, which might be
malicious, and this addition can overflow. Moreover, adding size_t might
cause overflows too on 32-bit platforms (such as WebAssembly), where
size_t == uint32_t.

If an oveflow happens here because the signature length field has been
corrupted just the right way, this length check on the buffer might
actually succeed, and allow a buffer overrun happen later during
signature computation.

In order to fix this condition, we need to avoid the overflow. uint64_t
should be enough to store the total expected length of a signed message
because uint32_t + uint32_t + 12 (the size of header as size_t) will fit
into uint64_t without causing overflows.

This whole issue could have been avoided if Secure Message in signature
mode produced the same length field as in the encryption mode: total
length including the header. However, it does not, and hence this fix.