Remove misleading length checks in decrypt/verify
Okay, here's how Secure Messages look like.

For encrypt/decrypt mode:

        4 bytes      4 bytes     variable
    +------------+------------+-----------+
    |  msg type  | msg_length | (payload) |
    +------------+------------+-----------+

For sign/verify mode:

        4 bytes      4 bytes      4 bytes     var.      variable
    +------------+------------+------------+--------+-------------+
    |  msg type  |  text len. | sign. len. | (text) | (signature) |
    +------------+------------+------------+--------+-------------+

Due to historical reasons, in encrypt/decrypt mode the "msg_length"
field contains the total length of Secure Message, including the payload
length and and the header length. However, in sign/verify mode there are
two length fields, and neither of them include the length of the header.
Instead, the total length of a signed Secure Message is the sum of those
lengths PLUS the 12 bytes of the header.

Now, THEMIS_SECURE_MESSAGE_LENGTH() does NOT return the total length of
Secure Message based on its type. It returns an uint32_t at the offset 4
which *is* the total length for encrypted messages, but for signed
messages that's just the length of the plaintext.

While it's technically okay to check that the total buffer size has at
least this length, it is very confusing and communicates an incorrect
message that this check concludes the length verification. This is
especially confusing in themis_secure_message_unwrap() which deals with
both types of Secure Messages (historical deprecated API, yada-yada).

Remove those checks to avoid confusion. The length of Secure Message
will be verified later by the themis_secure_message_verifier_proceed(),
or the themis_secure_message_ec_decrypter_proceed(),
or the themis_secure_message_rsa_decrypter_proceed(), after they have
determined what kind of Secure Message we are dealing with now.