Add more pom.xml metadata
Currently we're defining the bare minimum: artifact name and version.
However, Maven Central repository wants more information about the
artifacts.

This is a bit different from what we include on Bintray, but it's
consistent with pom.xml manually published on Maven Central,
and *that* is consistent with other language wrappers.