Mention migration in changelog
As you might know, JFrog is sunsetting Bintray and JCenter [1].
Unfortunately, this is exactly where JavaThemis has been hosted so far.
In order to ensure continuous availability, JavaThemis migrates to the
Maven Central repository [2].

[1]: https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/
[2]: https://maven.apache.org/repository/index.html

Maven Central repository is maintained by the Apache Software Foundation.
It is much less likely to experience a critical existence failure in the
future. Maven Central repository is well established and trusted in the
developer community.

Gradle has built-in support for the Central repository. Many Android
apps are probably already using it. The only change that app developers
need to do is slightly tweak the build.gradle file:

    repositories {
        // NEW repository:
        mavenCentral()

        // Old and deprecated:
        // jcenter()
    }

    // NOTE: this section is exactly the same as it was for JCenter
    dependencies {
        // Add JavaThemis as runtime dependency of your application.
        // Always pin the latest version, you can find it here:
        // https://search.maven.org/artifact/com.cossacklabs.com/themis
        implementation 'com.cossacklabs.com:themis:0.13.1'
    }

Latest stable release of Themis 0.13.1 is already published in the Maven
Central repository (as well as available via Bintray and JCentral, until
they are discontinued). New releases will be published only in the Maven
Central repository.