Publish JavaThemis for Android to Maven Central (#786)
* Add more pom.xml metadata

Currently we're defining the bare minimum: artifact name and version.
However, Maven Central repository wants more information about the
artifacts.

This is a bit different from what we include on Bintray, but it's
consistent with pom.xml manually published on Maven Central,
and *that* is consistent with other language wrappers.

* Add Sonatype staging repository

Define the repository we will be publishing to: Sonatype staging
repository. First artifacts go there, then they are transferred to
the OSS repository from which they are mirrored to the Central repo.

* Sign artifacts before publishing

Here's this Gradle plugin which everyone uses. Configure it to sign the
published artifacts with GPG. I don't want to put keys into some weird
Java thingie so let the system's GPG manage them.

The "signing" plugin has pretty limited configurability. Most of it
should be set in "gradle.properties" file in your home directory.

* Remove Bintray publishing

Now that we're ready to publish to Maven Central repo, there is no
longer a need to keep Bintray publishing toolchain around. Drop it.
You have served us well but publishing to the central repo turned out
to be not *that* onerous.

* Update publishing task exception list

We use a vile hack to add BoringSSL as a dependency for AndroidThemis
and this hack needs all publishing-related tasks to be enumerated for
them to not automatically apply to BoringSSL which we don't want to
publish.

* Update comments about "org.gradle.configureondemand"

Turns out that it was Bintray plugin that was broken with this option.
Publishing to standard Maven repos is fine.

* Update comments and add TODOs

There are a couple of places which could use improvements. Note them in
TODO comments to acknowledge the weirdness of whatever we're doing here.

* Mention migration in changelog

As you might know, JFrog is sunsetting Bintray and JCenter [1].
Unfortunately, this is exactly where JavaThemis has been hosted so far.
In order to ensure continuous availability, JavaThemis migrates to the
Maven Central repository [2].

[1]: https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/
[2]: https://maven.apache.org/repository/index.html

Maven Central repository is maintained by the Apache Software Foundation.
It is much less likely to experience a critical existence failure in the
future. Maven Central repository is well established and trusted in the
developer community.

Gradle has built-in support for the Central repository. Many Android
apps are probably already using it. The only change that app developers
need to do is slightly tweak the build.gradle file:

    repositories {
        // NEW repository:
        mavenCentral()

        // Old and deprecated:
        // jcenter()
    }

    // NOTE: this section is exactly the same as it was for JCenter
    dependencies {
        // Add JavaThemis as runtime dependency of your application.
        // Always pin the latest version, you can find it here:
        // https://search.maven.org/artifact/com.cossacklabs.com/themis
        implementation 'com.cossacklabs.com:themis:0.13.1'
    }

Latest stable release of Themis 0.13.1 is already published in the Maven
Central repository (as well as available via Bintray and JCentral, until
they are discontinued). New releases will be published only in the Maven
Central repository.