Maven publishing for desktop JavaThemis
Let's put JavaThemis for desktop systems into Maven Central as well.

Note that this includes only the Java part of the library -- a JAR with
JVM bytecode. JavaThemis also needs a JNI library, which you will have
to install separately, as appropriate for your system. See docs:

https://docs.cossacklabs.com/themis/languages/java/installation-desktop/

Maven Central requires Javadocs and source code to be available so make
them available. The tasks to build them are almost the same as for
Android, but not quite the same. Building source JAR is the most funny
one as it has uncovered an issue in Gradle which was not visible when
building Android JARs (as that happens in a different directory).

Also note that we are publishing under the same "weird" groupId:
"com.cossacklabs.com", because that's what we have used in the past, and
that's our groupId prefix on Maven Central.

Since the artifactId "themis" is already taken by JavaThemis for Android,
desktop systems will use "java-themis". Hopefully, this will not be too
confusing for the users.