Maven publishing for desktop JavaThemis (#788)
* Maven publishing for desktop JavaThemis

Let's put JavaThemis for desktop systems into Maven Central as well.

Note that this includes only the Java part of the library -- a JAR with
JVM bytecode. JavaThemis also needs a JNI library, which you will have
to install separately, as appropriate for your system. See docs:

https://docs.cossacklabs.com/themis/languages/java/installation-desktop/

Maven Central requires Javadocs and source code to be available so make
them available. The tasks to build them are almost the same as for
Android, but not quite the same. Building source JAR is the most funny
one as it has uncovered an issue in Gradle which was not visible when
building Android JARs (as that happens in a different directory).

Also note that we are publishing under the same "weird" groupId:
"com.cossacklabs.com", because that's what we have used in the past, and
that's our groupId prefix on Maven Central.

Since the artifactId "themis" is already taken by JavaThemis for Android,
desktop systems will use "java-themis". Hopefully, this will not be too
confusing for the users.

* Set up publishing to Sonatype staging repos

Configure the repositories and signing keys. Those should the be same
as used for publishing Android binaries and they should be set up in
the same way.

* Include desktop systems into the changelog

Add this PR to the changelog, make the wording more inclusive, and
provide some copypastable examples of "build.gradle" configurations.

Publishing JavaThemis for desktop in *any* Maven repository is a new
thing. Android publishings were long available via JCenter and Bintray,
but JavaThemis for desktop has never been published anywhere. Until this
moment, that is.

* Remove unnecessary plugins for Bintray publishing

We don't publish to Bintray anymore, hence those Gradle plugins are
no longer necessary. Drop them while we are here.