Test MSYS2 build and packaging
The documentation claims that we regularly test MSYS2 packaging but
actually never do. Start doing that.

Since MSYS2 is *very* different from Linux/macOS, make a separate task
for it so that we can override the shell to MSYS2 shell. While we're at
it, test everything that we have:

  - plain build in MSYS2 environment and runnin unit tests
  - building and installing packages, then testing the examples
  - building NSIS installer

Unfortunately, I don't really have a good idea of how to test the
NSIS installer, which will probably require a Visual Studio project
or something like that. But it would be cool to check if it works too.

Also, to avoid spending more time on this than necessary, test only
OpenSSL builds with system packages provided by MSYS2, don't test
building BoringSSL and appropriate Themis flavor.