Update zeroize to "1" (#799)
This is the only dependency of RustThemis and they have managed to break
even that. Some bright maintainer head over there thought that it would
be a great idea to just yank all 0.x versions from crates.io, breaking
our build. (Since RustThemis is a library, we're not using Cargo.lock.
If a dependency is yanked, we can't build our stuff.)

Well, thanks for not breaking the API in 1.x at least. You're still
better than the Tokio team.

Note that zeroize requires reasonably new stable version of Rust
toolchain. With 0.5.2 this has been 1.31, now it's 1.47. (The current
stable version is 1.51. RustThemis tracks that version.)