New test unit-targets for BoringSSL-based builds
Let's test the new builds. Add a couple of new test targets that are
going to exercise BoringSSL-based builds.

Use only Swift 5 version of the target. It's enough to test Swift 4 with
only OpenSSL (while it exists). The purpose of this test is to check
Themis API for Swift, not the backend.

Note that *these* targets *are* shared. We are going to run those on CI,
for example, so they need to be present in the checkout.