Build BoringSSL with Xcode (#743)
* BoringSSL intermediate file generator

Start off by adding a utility script to generate intermediate files
required by BoringSSL build. Normally, this step is performed by the
build system configuration (e.g., CMake). It requires a number of
utilities such as Perl and Go to generate the error list and all the
assembly code. Once this is done, the intemediate files will be checked
into the repository and reused.

This script will be executed as part of Make rule for updating BoringSSL.
It is not expected to be used manually, but you can certainly do this.

Note that BoringSSL's "util/generate_build_files.py" script is pretty
picky. It expects the BoringSSL source tree to be located in "src"
directory, it works only with Python 2, and it spews out generated files
into the current directory and expects them to stay there.

Also note that we save not only the generated source files but also
the "boringssl.gypi" file. Our build system does not use GYP, but its
project file is the easiest one to parse. We'll use it later to get the
list of all BoringSSL source files to use.

* BoringSSL intermediate files

Run "third_party/boringssl/generate_build_files.sh" and check in the
generated files.

* Build BoringSSL with Xcode

Add "BoringSSL" group to Xcode project, currently without any files.
Also add "BoringSSL (iOS)" and "BoringSSL (macOS)" targets building
static libraries "libboringssl.a" for iOS and macOS respectively.

These are currently placeholder and next commits in this pull request
will fill them in with files to compile and adjust compilation settings.

* Script for updating BoringSSL targets of Xcode project

BoringSSL does not have a native build system for iOS *and* macOS. It is
possible to coerce CMake into generating some Xcode project which might
work, but it requires installing custom toolchains, and it's not very
reliable nor convenient for us. Therefore, we take the matter into our
own hands and update existing Xcode project manually, using the
intermediate build files generated earlier.

The "third_party/boringssl/update_xcodeproj.rb" script will read the GYP
file, extract the source file lists, and update Themis.xcodeproj to use
those source files appropriately. Files will be added to the "BoringSSL"
file group, the source files will be added to the compilation steps,
this includes both C/C++ source code and assembly as well.

We're using CocoaPods' "xcodeproj" gem and write the script in Ruby
because that's the most stable and battle-tested programmatic way to
manipulate Xcode projects.

* Update Xcode project from BoringSSL's GYP file

Run the "third_party/boringssl/update_xcodeproj.rb" script and update
the Xcode project with BoringSSL file list.

* Configure BoringSSL targets

Adjust Xcode project configuration for building BoringSSL.

  - Build "arm64e" architecture for iOS, like we do for Themis.

  - Build assembly files only for relevant architectures.

    This one is very tricky to do and works only due to pure luck.

    Xcode does not really offer a convenient way to include/exclude
    individual files based on architecture. However, there are two
    settings which more or less allow it: EXCLUDED_SOURCE_FILE_NAMES
    and INCLUDED_SOURCE_FILE_NAMES.

    We set up those settings to exclude the assembly files by default,
    but include them back for relevant SDKs and architectures. Real
    iOS devices use ARM64 (and ARMv7 for iOS 10 which we still support)
    so they need the ARM assembly stuff. macOS devices use x86_64 but
    can run x86 apps (until 10.15), so we include both x86 and x86_64
    assembly stuff. New Apple Silicon devices will use ARM64, but
    BoringSSL does not include ARM assembly for macOS yet (and we can't
    blindly use the one designed for iOS devices).

    iOS Simulator does not simulate any hardware acceleration so we add
    OPENSSL_NO_ASM preprocessor definiton to block assembly usage in
    BoringSSL build and use pure software implementation there.

    Similarly, we currently use OPENSSL_NO_ASM for arm64 on macOS too
    because BoringSSL build expects hardware acceleration, but we can't
    provide it (yet).

  - Disable compilation warnings.

    Since BoringSSL is a third-party project, we don't have any direct
    control over its code and can't fix the warnings it produces. And it
    produces quite a few. Inhibit all warnings there and don't bother.

* New make target: update_boringssl

Add a utility target "update_boringssl" to update BoringSSL: pull the
latest source code into the submodule and run all the update scripts.

Xcode project editing scripts are run only when running on macOS since
that's required to test the updates, and likely to perform them in the
first place.

* Update to latest BoringSSL

Run the "make update_boringssl" target and update to latest BoringSSL.
This includes both the submodule update and the Xcode build files.

* Remove "Recovered references" after merge

Merging Xcode's project XMLs is so much fun! It's easier to just redo
the work on top of the updated project.

* Use whatever Python is available in the system

Recent versions of BoringSSL are now able to work with either Python 3
or Python 2, so we don't particularly care about that anymore. Just
launch *some* Python and we're good.

* Update to latest BoringSSL (once again)

Run "make update_boringssl", commit whatever the updates are.

* Remove "arm64e" architecture from iOS builds

Just build with the standard architecture set. Don't try building arm64e
which tends to cause multiple arcane issues without providing any value.
Apple has been recommending to "test your builds" with it, but it has
never recommended deploying applications that use it. So screw it.

* Include ARM assembly for macOS as well

Try using ARM assembly files generated for iOS during macOS builds too.
I don't have an M1 laptop so I have no clue if it actually works, but it
certainly does compile. Maybe it does not use those fast paths, but vOv

Add those new files into Xcode project.

* Actually include ARM files into macOS builds

Update the source file exclusion/inclusion rules to build ARM64 assembly
files for macOS builds.

* Don't set OPENSSL_NO_ASM for macOS anymore

Now that we actually have some assembly in there, try using it. I'm not
sure that it's actually being used though, too lazy to disassmeble.

* Adjust changelog to mention only BoringSSL update

Since it's likely that BoringSSL-based builds of Themis won't make it
into the 0.14.0 release, let's not put more that we can eat onto the
plate. This patch set is going to just add BoringSSL targets to the
Xcode project and a new "update_boringssl" target for Make, that's it.

* Remove BoringSSL mentions from Objective-C and Swift sections

Since there is no user-visible impact of these changes on Objective-C
and Swift users, there is no point in adding those lines there, right?