Copy cossacklabs/themis-java-examples/android-example
Start by importing the example files. While I could import the history
as well, there is not much value in that, so here are just the files.

Note that this example uses Gradle, like Themis project itself, but the
example is completely independent of the top-level Gradle project. It
even uses a different version of Gradle wrapper because that's what
Android Studio insists on. I'd like to keep it that way, so that users
are able to copy and paste the project directory and get a working
example, not dependent on any Themis structure.

The example is also tied to a published version of AndroidThemis, not
the one being developed in the source tree. Maintainers will have to
update the dependency manually, similar to how macOS/iOS examples work.