Assimilate Android example project (#813)
* Copy cossacklabs/themis-java-examples/android-example

Start by importing the example files. While I could import the history
as well, there is not much value in that, so here are just the files.

Note that this example uses Gradle, like Themis project itself, but the
example is completely independent of the top-level Gradle project. It
even uses a different version of Gradle wrapper because that's what
Android Studio insists on. I'd like to keep it that way, so that users
are able to copy and paste the project directory and get a working
example, not dependent on any Themis structure.

The example is also tied to a published version of AndroidThemis, not
the one being developed in the source tree. Maintainers will have to
update the dependency manually, similar to how macOS/iOS examples work.

* Bump Gradle version 4.1.2 => 4.1.3

Android Studio suggests this change, let's go with it. I guess we could
make a bit of a wider jump to get to the current 6.x versions, but maybe
a bit later.

* Bump target SDK 26 => 29

* Migrate from "android.support" library to "androidx.appcompat"

Another sacrifice to the CADT model because Android Studio is so
insistent on it. At least they have "Refactoring | Migrate to AndoridX"
which has helped a lot.

* Bump androidx.appcompat 1.0.0 => 1.2.0

* Bump androidx.constraintlayout 1.1.3 => 2.0.4

* Configure Kotlin for the project

I'm using 1.3.72 because that's what my Android Studio comes with and
it's pretty insistent on using that. The latest version is 1.5.0.

* Convert Android example to Kotlin

Automatic conversion with "Code | Convert Java file to Kotlin file".
This works reasonably well for initial conversion, but more work needs
to be done manually to fix up some remaining bits.

I'm converting Android example to Kotlin because that's how most Android
stuff is developed these days. Java is really a legacy language there.
Java 16 is the current version. Android has mostly Java 7 with some
selected bits of Java 8. Go figure.

JavaThemis will have its own separate example using Java specifically.
AndroidThemis now has a Themis showcase.

* Adjust nullability here and there

Automatic conversion did not get the fact that ISessionCallbacks has
nullable return types. Maybe that's because they don't have annotation?

* Bump Kotlin 1.3.72 => 1.4.32

Android Studio has suddenly discovered that it has some update. Gotta
update then, eh.

* Use Kotlin-style array comparisons

Now it's a flurry of miscellaneous IDE-suggested updates.

* Don't use properties where parameters suffice

* Relax exception annotations

Kotlin is not very keen on Java checked exceptions. All exceptions there
are unchecked, runtime ones. Drop all the exception annotations and
fine-grained catch clauses, replace them with catch-alls. Currently we
don't do any involved exception handling, we just want to prevent app
crashes. Specific exception handling will have specific exceptions.

* Explicit Themis imports

Make sure we import everything we need explicitly. Previously we have
also imported a bunch of exception types which we now don't need.

* Build Android examples on GitHub Actions

Just build. It would be cool to test the resulting application but that
opens up a whole new time hell hole, with launching emulators and stuff.
I have no desire to dive there at the moment.

* Update changelog

This is somewhat notable so let's update the changelog too. Once
JavaThemis example is moved to the main repository as well, we can
finally deprecate and archive that separate repo.