Build Themis + BoringSSL with Xcode (#810)
* Add "src/soter/boringssl" into Xcode project

Start by adding the source files we'll need for compilation. And don't
ask me what are all those other changes in there, I have no clue.

* Xcode targets for building Themis with BoringSSL

A couple new targets will build Themis with BoringSSL as opposed to
Carthage-provided OpenSSL:

  - Themis (iOS) - BoringSSL
  - Themis (macOS) - BoringSSL

Note that these come *without* shared schemes. This is to prevent
Carthage from building them and overwriting OpenSSL-based builds.
The targets build the same framework -- "themis.framework" -- and
Carthage is too stupid to put results into separate directories or
something.

* Build BoringSSL variant of Soter for BoringSSL builds

Replace "src/soter/openssl/*" with "src/soter/boringssl/*" in source
file lists for BoringSSL-based builds.

* Use embedded BoringSSL headers

Make sure that we can locate BoringSSL headers during builds.

* Configure BoringSSL preprocessor definitions

Define some things that configure Themis for BoringSSL. Normally these
are defined by the Makefile, but we don't have that in Xcode, obviously.

* Link against embedded BoringSSL

Replace OpenSSL linkage with BoringSSL.

* New test unit-targets for BoringSSL-based builds

Let's test the new builds. Add a couple of new test targets that are
going to exercise BoringSSL-based builds.

Use only Swift 5 version of the target. It's enough to test Swift 4 with
only OpenSSL (while it exists). The purpose of this test is to check
Themis API for Swift, not the backend.

Note that *these* targets *are* shared. We are going to run those on CI,
for example, so they need to be present in the checkout.

* Link BoringSSL-based tests with BoringSSL-based build

Replace OpenSSL-based builds of Themis with BoringSSL-based ones in
dependencies and linkage instructions for the new unit-test targets.

* Exercise BoringSSL unit tests on GitHub Actions

Piggyback on the "Unit tests (Carthage)" job to run the tests. We don't
really need Carthage here, but that's where unit tests from the main
Xcode project are being run.

Make sure to check out BoringSSL submodule because we're obviously going
to compile it in new tests.