Unbreak Android build (#824)
* Uninstall CMake 3.18.1 from Android SDK

Gradle ends up really confused by this CMake version:

    FAILURE: Build failed with an exception.

    * What went wrong:
    Could not determine the dependencies of task ':android:assembleDebug'.
    > A problem occurred configuring project ':boringssl'.
       > Invalid revision: 3.18.1-g262b901

Our builds are fine with older CMake (for now), so let's use that.

* Pin runner to macos-10.15

GitHub people are working on adding macOS 11 support for Actions and
it's probably going to happen soon. However, I don't really want to be
notified about new and exciting ways to break our builds when that
finally happens, so let's keep using the old environment for now.