BoringSSL support for XCFramework builds (#825)
* Improve readability with newlines and indents
* Remove boilerplate from comments
* Check working directory in scripts/create_xcframeworks.sh
* Don't save PWD, use "cd ~-" instead

We have to change the directory only because that's how "zip" works.
Don't make this hack infect the rest of the script.

* Pull Carthage dependencies on first run in clear directory
* Split XCFramework build into functions
* Accept "--with-clopenssl" and "--with-boringssl" options

Initially recognize "--with-clopenssl" and "--with-boringssl" to control
what kind of Themis flavor we're building. (Maybe one, maybe both.)

If nothing is specified, build CLOpenSSL-based build, like before.

* Do "carthage bootstrap" only if building with CLOpenSSL
* Support "--help" option too

I'm always up for inventing new silly ways to do self-descripting
implementations of "--help". This one is an evolution of my previous
approach of reading the header of the script, now with variable
expansion so that "$BUILD_PATH" is actually displayed there.

* Make "output_dir" a parameter of helper functions
* Move CLOpenSSL output into "build/xcf_output/CLOpenSSL"

Make way for side-by-side BoringSSL builds. Both are going to produce
the same "themis.xcframework", but you can't have two files with the
same name in the same directory, don't you? They're like fermions.

* Make Xcode schemes parameterizable too

This is where we're getting close to supporting BoringSSL builds too.
THey are using different Xcode schemes to produce the same frameworks.

* Actually support building both CLOpenSSL and BoringSSL flavors

Respect the flags set on the command line and build the requested
flavors. Note how we interleave the build steps, making sure to first
build the binaries, then pack them, then compute checksums. This makes
output of this script easier to check. What you're really interested
is whether everything has been included and what are the checksums.
Build logs are only there in case they break.

* Divide the build actions into sections more cleanly
* Print filenames for XCFramework checksums

Now that we can build two different XCFrameworks, be clear what exactly
we're checksumming. This also serves as a reminder of where to find
build outputs.